refactor: simplify builder code
