+++ packages/shared/src/devServer.ts
@@ -1,20 +1,16 @@
 import type {
   RsbuildConfig,
   StartDevServerOptions,
-  StartServerResult,
   Context,
   OnAfterStartDevServerFn,
   OnBeforeStartDevServerFn,
   CompilerTapFn,
   DevServerOptions,
-  ServerApi,
 } from './types';
-import { getAddressUrls } from './url';
-import { isFunction } from './utils';
 import { getPort } from './port';
 import deepmerge from 'deepmerge';
 import { color } from './color';
-import { logger as defaultLogger, debug, Logger } from './logger';
+import { logger as defaultLogger, Logger } from './logger';
 import { DEFAULT_PORT, DEFAULT_DEV_HOST } from './constants';
 import { createAsyncHook } from './createHook';
 import { mergeChainedOptions } from './mergeChainedOptions';
@@ -42,6 +38,41 @@ export function printServerURLs(
   logger.log(message);
 }
 
+export const getDevOptions = async ({
+  rsbuildConfig,
+  serverOptions,
+  strictPort,
+  getPortSilently,
+}: {
+  rsbuildConfig: RsbuildConfig;
+  serverOptions: ServerOptions;
+  strictPort: boolean;
+  getPortSilently?: boolean;
+}) => {
+  const port = await getPort(rsbuildConfig.dev?.port || DEFAULT_PORT, {
+    strictPort,
+    silent: getPortSilently,
+  });
+
+  const host =
+    typeof serverOptions?.dev === 'object' && serverOptions?.dev?.host
+      ? serverOptions?.dev?.host
+      : DEFAULT_DEV_HOST;
+
+  const https =
+    typeof serverOptions?.dev === 'object' && serverOptions?.dev?.https
+      ? Boolean(serverOptions?.dev?.https)
+      : false;
+
+  const devServerConfig = getDevServerOptions({
+    rsbuildConfig,
+    serverOptions,
+    port,
+  });
+
+  return { port, host, https, devServerConfig };
+};
+
 export const getDevServerOptions = async ({
   rsbuildConfig,
   serverOptions,
@@ -50,9 +81,7 @@ export const getDevServerOptions = async ({
   rsbuildConfig: RsbuildConfig;
   serverOptions: ServerOptions;
   port: number;
-}): Promise<{
-  devConfig: DevServerOptions;
-}> => {
+}): Promise<DevServerOptions> => {
   const defaultDevConfig = deepmerge(
     {
       hot: rsbuildConfig.dev?.hmr ?? true,
@@ -71,13 +100,11 @@ export const getDevServerOptions = async ({
     (serverOptions.dev as Exclude<typeof serverOptions.dev, boolean>) || {},
   );
 
-  const devConfig = mergeChainedOptions({
+  return mergeChainedOptions({
     defaults: defaultDevConfig,
     options: rsbuildConfig.tools?.devServer,
     mergeFn: deepmerge,
   });
-
-  return { devConfig };
 };
 
 /** The context used by startDevServer. */
@@ -92,106 +119,6 @@ export type DevServerContext = Context & {
   };
   config: Readonly<RsbuildConfig>;
 };
-
-export async function startDevServer<
-  Options extends {
-    context: DevServerContext;
-  },
->(
-  options: Options,
-  createDevServer: (
-    options: Options,
-    port: number,
-    serverOptions: ServerOptions,
-    compiler: StartDevServerOptions['compiler'],
-  ) => Promise<ServerApi>,
-  {
-    open,
-    compiler,
-    printURLs = true,
-    strictPort = false,
-    serverOptions = {},
-    logger: customLogger,
-    getPortSilently,
-  }: StartDevServerOptions & {
-    defaultPort?: number;
-  } = {},
-) {
-  const logger = customLogger ?? defaultLogger;
-
-  if (!process.env.NODE_ENV) {
-    process.env.NODE_ENV = 'development';
-  }
-
-  const rsbuildConfig = options.context.config;
-
-  const port = await getPort(rsbuildConfig.dev?.port || DEFAULT_PORT, {
-    strictPort,
-    silent: getPortSilently,
-  });
-
-  const host =
-    typeof serverOptions?.dev === 'object' && serverOptions?.dev?.host
-      ? serverOptions?.dev?.host
-      : DEFAULT_DEV_HOST;
-
-  const https =
-    typeof serverOptions?.dev === 'object' && serverOptions?.dev?.https
-      ? Boolean(serverOptions?.dev?.https)
-      : false;
-
-  options.context.devServer = {
-    hostname: host,
-    port,
-    https,
-    open,
-  };
-
-  const protocol = https ? 'https' : 'http';
-  let urls = getAddressUrls(protocol, port, rsbuildConfig.dev?.host);
-
-  if (printURLs) {
-    if (isFunction(printURLs)) {
-      urls = printURLs(urls);
-
-      if (!Array.isArray(urls)) {
-        throw new Error('Please return an array in the `printURLs` function.');
-      }
-    }
-
-    printServerURLs(urls, logger);
-  }
-
-  const server = await createDevServer(options, port, serverOptions, compiler);
-
-  await options.context.hooks.onBeforeStartDevServerHook.call();
-
-  debug('listen dev server');
-  await server.init();
-
-  return new Promise<StartServerResult>((resolve) => {
-    server.listen(
-      {
-        host,
-        port,
-      },
-      async (err: Error) => {
-        if (err) {
-          throw err;
-        }
-
-        debug('listen dev server done');
-
-        await options.context.hooks.onAfterStartDevServerHook.call({ port });
-        resolve({
-          port,
-          urls: urls.map((item) => item.url),
-          server,
-        });
-      },
-    );
-  });
-}
 
 type ServerCallbacks = {
   onInvalid: () => void;
@@ -215,9 +142,9 @@ export const setupServerHooks = (
 
   const { compile, invalid, done } = compiler.hooks;
 
-  compile.tap('modern-dev-server', hookCallbacks.onInvalid);
-  invalid.tap('modern-dev-server', hookCallbacks.onInvalid);
-  done.tap('modern-dev-server', hookCallbacks.onDone);
+  compile.tap('rsbuild-dev-server', hookCallbacks.onInvalid);
+  invalid.tap('rsbuild-dev-server', hookCallbacks.onInvalid);
+  done.tap('rsbuild-dev-server', hookCallbacks.onDone);
 };
 
 export const isClientCompiler = (compiler: {
